# perf: replace `commander` with `cac`