--- a/packages/core/rslib.config.ts
+++ b/packages/core/rslib.config.ts
@@ -50,7 +50,6 @@ export default defineConfig({
   output: {
     externals: {
       picocolors: '../compiled/picocolors/index.js',
-      commander: '../compiled/commander/index.js',
       chokidar: '../compiled/chokidar/index.js',
       rslog: '../compiled/rslog/index.js',
     },