--- a/packages/core/tsconfig.json
+++ b/packages/core/tsconfig.json
@@ -11,7 +11,6 @@
     "module": "ESNext",
     "moduleResolution": "Bundler",
     "paths": {
-      "commander": ["./compiled/commander"],
       "chokidar": ["./compiled/chokidar"],
       "picocolors": ["./compiled/picocolors"],
       "rslog": ["./compiled/rslog"]