… packages/core/package.json
… pnpm-lock.yaml