## Summary

Replace `commander` with `cac` to improve cli performance.

### Output Size

`cac` is 18kB smaller than `commander`

- cac:

```
File (esm)                    Size       
dist/index.js                 130.4 kB
```

- commander:

```
File (esm)                    Size       
dist/index.js                 111.9 kB
```

### Startup Time

`cac` is 11ms faster than `commander`.

- cac:

```
hyperfine --warmup 5 'node rslib.js build'
Benchmark 1: node /Users/bytedance/codes/rslib/packages/core/bin/rslib.js build
  Time (mean ± σ):      1.169 s ±  0.022 s    [User: 3.263 s, System: 0.324 s]
  Range (min … max):    1.135 s …  1.198 s    10 runs
```

- commander:

```
hyperfine --warmup 5 'node rslib.js build'
Benchmark 1: node /Users/bytedance/codes/rslib/packages/core/bin/rslib.js build
  Time (mean ± σ):      1.148 s ±  0.023 s    [User: 3.259 s, System: 0.312 s]
  Range (min … max):    1.121 s …  1.188 s    10 runs
```

## Related Links

https://github.com/cacjs/cac

<!--- Provide links of related issues or pages -->

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [ ] Documentation updated (or not required).